# Optimize TypedSet and map_concat, array_union