… presto-main/src/main/java/com/facebook/presto/operator/scalar/MapConcatFunction.java
… presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestOptimizedTypedSet.java