This PR resolves the "Avoid block building inside" part in https://github.com/prestodb/presto/issues/15361. The optimizations included in this PR are
- Avoid block building inside.

  TypedSet has an internal BlockBuilder and appends each Block position being added to it. Block building using BlockBuilder is highly costly and inefficient operation. Here the BlockBuilder is needed to 1) resolve hash table probing collision 2) rehash. However, both are actually not a problem. In most use cases, whole blocks (instead of several positions of a block) are added to the TypedSet, and problem 1) can be resolved by keeping track of the blocks being added. 2) Rehashing is not needed since we can know the max number of entries before creating a TypedSet for most use cases. So we want to provide a method that adds a whole Block and just record the positions in the set using SelectedPositions objects. By providing this new interface, internal operations can be streamlined to more efficient loops. It also gives opportunity to get the result Block using more efficient APIs that allows encapsulated memory copying.

  In the new OptimizedTypedSet class, we offer several operations: union, intersect and except. The operations can be called multiple times, but user does need to specify the max set size when creating the OptimizedTypedSet.

- Avoid computing the hash positions multiple times
The previous TypedSet usage sometimes require calculating the hash position multiple times if there are multiple TypedSets. For example array_intersect and array_except. Such usages build one TypedSet R, and based on the probe result on R, insert new elements to another TypedSet B. It requires to calculate the hash position multiple times. This can be avoided if the new TypdedSet can encapsulate the operations inside. The hashPosition calculated in hashtable A can be reused by hashtable B if the size and hash functions are the same.

- JMH benchmark shows up to 40% improvement for `array_union`:

    Type       | Baseline  | Specialized Baseline | OptimizedTypedSet | Gain%
    -----------|-----------|----------------------|-------------------|----------
    BIGINT     |      5511 |                 3742 |            3320   |      40%
    VARCHAR    |     20414 |                 N/A  |            14155  |      31%

- JMH benchmark shows 82% improvement for non_empty case for `map_concat` when keyCount=100
    and POSITIONS=1000:

```
    Baseline

    Benchmark                     Mode  Cnt      Score      Error  Units
    BenchmarkMapConcat.mapConcat  avgt   20  26710.925 ± 2005.756  ns/op
    Retained Size: 1,402,374 bytes

    After
    Benchmark                     Mode  Cnt      Score      Error  Units
    BenchmarkMapConcat.mapConcat  avgt   20  14605.437 ± 1209.786  ns/op
    Retained Size: 1,373,273 bytes

```
 When keyCount=1000 and POSITIONS=1000, **the baseline just OOMed. The optimized version succeeded.**

- Array_intersect  showed up to 49% improvement in time and 72% savings in allocation rate. More detailed comparisons:
![Screen Shot 2020-11-05 at 1 59 55 AM](https://user-images.githubusercontent.com/33299678/98228163-14447a00-1f0d-11eb-9389-20df2a1ae3c6.png)
![Screen Shot 2020-11-05 at 2 00 02 AM](https://user-images.githubusercontent.com/33299678/98228174-173f6a80-1f0d-11eb-8ba0-a86cbdda2c6b.png)

- array_except JMH benchmark shows 40% improvement:

```
    Before
    Benchmark                               Mode  Cnt       Score        Error  Units
    BenchmarkArrayIntersect.arrayIntersect  avgt   10  618074.452 ± 119912.203  ns/op

    After
    Benchmark                               Mode  Cnt       Score       Error  Units
    BenchmarkArrayIntersect.arrayIntersect  avgt   10  376854.064 ± 21616.063  ns/op
```
Production testing:
Tested using verifier on 1220 queries with map_concat, array_union, array_intersect, array_except. No failures found.

```
== NO RELEASE NOTE ==
```