@kaikalur 2020-10-27
The perf numbers look good. But is there a memory overhead? Is there a way to measure that?