@yingsu00 2020-10-28
> The perf numbers look good. But is there a memory overhead? Is there a way to measure that?

@kaikalur Sreeni, thank you very much for reviewing the PR. Let's see the top memory consumers in OptimizedTypedSet

- Block[] blocks 
This is to keep reference of the incoming blocks. In original implementation, the caller of TypedSet also need to keep track of them implicitly while adding elements to it. This not increasing memory consumption nor prolonging Block object life cycle.

- List<SelectedPositions> positionsForBlocks
The list of SelectedPositions is a small fraction of the original block builder `elementBlock`. Since we removed `elementBlock` the memory actually dropped.

- int[] blockPositionByHash
The blockPositionByHash was the hash table which was the `IntArrayList blockPositionByHash` in original `TypedSet`. Since we have the upper limit of the hash table size, there is no need to use a dynamically growing structure. Using `int[]` is more efficient.

So look at the retained size calculation, it's now
```
public long getRetainedSizeInBytes()
    {
        long sizeInBytes = INSTANCE_SIZE + ARRAY_LIST_INSTANCE_SIZE + sizeOf(blockPositionByHash);
        for (int i = 0; i <= currentBlockIndex; i++) {
            sizeInBytes += sizeOf(positionsForBlocks.get(i).getPositions());
        }
        return sizeInBytes;
    }
```
vs. the one in TypedSet
```
public long getRetainedSizeInBytes()
    {
        return INSTANCE_SIZE + INT_ARRAY_LIST_INSTANCE_SIZE + elementBlock.getRetainedSizeInBytes() + blockPositionByHash.size() * Integer.BYTES;
    }
```
In BenchmarkMapConcat, the baseline TypeSet retained size was 530 bytes while the new implementation is 232 bytes.