@yingsu00 2020-10-28
@kaikalur This is the GC profile for BenchmarkMapConcat:

Baseline
```
Benchmark                                                      Mode  Cnt     Score      Error   Units
BenchmarkMapConcat.mapConcat                                   avgt   20  1090.895 ±   84.847   ns/op
BenchmarkMapConcat.mapConcat:·gc.alloc.rate                    avgt   20   692.512 ±   46.312  MB/sec
BenchmarkMapConcat.mapConcat:·gc.alloc.rate.norm               avgt   20  1565.082 ±   13.184    B/op
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Eden_Space           avgt   20   703.407 ±  633.867  MB/sec
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Eden_Space.norm      avgt   20  1606.711 ± 1448.248    B/op
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Survivor_Space       avgt   20     0.168 ±    0.285  MB/sec
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Survivor_Space.norm  avgt   20     0.377 ±    0.637    B/op
BenchmarkMapConcat.mapConcat:·gc.count                         avgt   20    10.000             counts
BenchmarkMapConcat.mapConcat:·gc.time                          avgt   20    14.000                 ms
```
Now:
```
Benchmark                                                      Mode  Cnt     Score      Error   Units
BenchmarkMapConcat.mapConcat                                   avgt   20   816.268 ±   71.719   ns/op
BenchmarkMapConcat.mapConcat:·gc.alloc.rate                    avgt   20   791.006 ±   58.960  MB/sec
BenchmarkMapConcat.mapConcat:·gc.alloc.rate.norm               avgt   20  1337.685 ±    0.005    B/op
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Eden_Space           avgt   20   739.708 ±  738.168  MB/sec
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Eden_Space.norm      avgt   20  1296.181 ± 1295.956    B/op
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Survivor_Space       avgt   20     0.111 ±    0.316  MB/sec
BenchmarkMapConcat.mapConcat:·gc.churn.PS_Survivor_Space.norm  avgt   20     0.208 ±    0.602    B/op
BenchmarkMapConcat.mapConcat:·gc.count                         avgt   20     9.000             counts
BenchmarkMapConcat.mapConcat:·gc.time                          avgt   20    13.000                 ms
```

The `gc.alloc.rate.norm`, `gc.churn.PS_Eden_Space.norm`, `gc.count`,`gc.time` all dropped.