@yingsu00 2020-11-05
> Couple of questions - I don't see NULL being handled explicitly anywhere. Hopefully we have some tests that take care of it.

@kaikalur Thanks for reviewing the PR. NULL was actually handled. In the past, there was a history where array functions didn't handle nulls correctly, please see https://github.com/prestodb/presto/issues/11978. It was because it was using `type.equalTo()` which couldn't distinguish 0 and null. This problem was corrected by using `positionEqualsPosition()` and the internal hashtable which can handle null correctly. The original logic of having a variable `hasNulls` was no longer needed.

In fact, all tests in `testOptimizedTypedSet` have cases for nulls. For example in `testUnionWithDistinctValues`
```
@Test
    public void testUnionWithDistinctValues()
    {
        OptimizedTypedSet typedSet = new OptimizedTypedSet(BIGINT, POSITIONS_PER_PAGE + 1, FUNCTION_NAME);

        Block block = createLongSequenceBlock(0, POSITIONS_PER_PAGE / 2);
        testUnion(typedSet, block, block);

        block = createLongSequenceBlock(0, POSITIONS_PER_PAGE);
        testUnion(typedSet, createLongSequenceBlock(POSITIONS_PER_PAGE / 2, POSITIONS_PER_PAGE), block);

        // Test adding a block with null
        Block blockWithNull = block.appendNull();
        testUnion(typedSet, blockWithNull, blockWithNull);
    }
```

To have even more test coverage, I added a new test in `TestOptimizedTypedSet`:
```
@Test
    public void testNulls()
    {
        OptimizedTypedSet typedSet = new OptimizedTypedSet(BIGINT, POSITIONS_PER_PAGE + 1, FUNCTION_NAME);

        // Empty block
        Block emptyBlock = createLongSequenceBlock(0, 0);

        testUnion(typedSet, emptyBlock, emptyBlock);

        // Block with a single null
        Block singleNullBlock = emptyBlock.appendNull();

        testUnion(typedSet, singleNullBlock, singleNullBlock);
        testIntersect(typedSet, singleNullBlock, singleNullBlock);
        testIntersect(typedSet, emptyBlock, emptyBlock);
        testExcept(typedSet, singleNullBlock, singleNullBlock);
        testIntersect(typedSet, emptyBlock, emptyBlock);

        // Block with a 0, and block with a 0 and a null
        Block zero = createLongSequenceBlock(0, 0);
        Block zeroAndNull = zero.appendNull();

        testUnion(typedSet, zero, zero);
        testUnion(typedSet, singleNullBlock, zeroAndNull);
        testIntersect(typedSet, zero, zero);
        testExcept(typedSet, singleNullBlock, singleNullBlock);
        testUnion(typedSet, zero, zeroAndNull);
    }
```